+++ src/sources/fast.rs
@@ -1,5 +1,7 @@
 use crate::sources::utils::de_currency;
 use crate::sources::{Currency, RateType, SourceCashUrlTrait};
+use rust_decimal::serde::arbitrary_precision;
+use rust_decimal::Decimal;
 use serde::Deserialize;
 
 pub const API_URL: &str =
@@ -27,13 +29,15 @@ impl SourceCashUrlTrait for Response {
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "PascalCase")]
 pub struct Rate {
-    pub buy: f64,
+    #[serde(deserialize_with = "arbitrary_precision::deserialize")]
+    pub buy: Decimal,
     #[serde(deserialize_with = "de_currency")]
     pub id: Currency,
     pub pay_type: u8,
     pub prev_buy: f64,
     pub prev_sale: f64,
-    pub sale: f64,
+    #[serde(deserialize_with = "arbitrary_precision::deserialize")]
+    pub sale: Decimal,
     pub sort_id: u64,
     pub unit: f32,
 }
